Fixed #72
